4.20 exercise corrected

+++ part4/bloglist/controllers/blogs.js
@@ -3,13 +3,6 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-// const getTokenFrom = request => {
-//   const authorization = request.get('authorization')
-//   if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-//     return authorization.substring(7)
-//   }
-//   return null
-// }
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
@@ -29,32 +22,38 @@ blogsRouter.get('/:id', async (request, response, next) => {
   }
 })
 
-blogsRouter.post('/', async (request, response, next) => {
-  const body = request.body
-  
-  const token = getTokenFrom(request)
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!token || !decodedToken.id) {
+blogsRouter.post('/', async (request, response) => {
+  try {
+    const decodedToken = await jwt.verify(request.token, process.env.SECRET)
+    request.decodedToken = decodedToken
+  } catch (error) {
+    request.decodedToken = null
+  }
+  if (!request.token || !request.decodedToken) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
-  const user = await User.findById(decodedToken.id)
+  const user = await User.findById(request.decodedToken.id)
 
   const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes === undefined ? 0 : body.likes,
+    title: request.body.title,
+    url: request.body.url,
+    likes: request.body.likes,
+    author: request.body.author,
     user: user._id
   })
 
-  try {
+  if (!blog.title || !blog.url || !blog.author) {
+    response.status(400).json({ error: 'incomplete submission' })
+  }
+  else {
     const savedBlog = await blog.save()
     user.blogs = user.blogs.concat(savedBlog._id)
+  
     await user.save()
-    response.json(savedBlog.toJSON())
-  } catch(exception) {
-    next(exception)
+  
+    response.status(201).json(savedBlog.toJSON())
   }
+
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
@@ -76,12 +75,25 @@ blogsRouter.put('/:id', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-  try {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+  const body = request.body
+  if (!body.token) {
+    console.log('not logged in')
   }
-  catch (exception) {
-    next(exception)
+  if (body.token) {
+    try {
+      console.log(request)
+      let hi = await Blog.findById(request.params.id)
+      if (hi.user == body.username) {
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+      }
+      else {
+        return response.status(401).json({ error: 'u not the user' })
+      }
+    }
+    catch (exception) {
+      next(exception)
+    }
   }
 })
